Add "I" Parameter to M305 For Software Control Of Heater Inversion
Useful for hardware setups which previously relied on the breakout of the heated bed MOSFET's gate pin, present on the 0.6 Duet, but removed in 0.8.5.

This is achieved by replacing references to the `HEAT_ON` constant with `Heat::GetHeatOn(heater)` in the appropriate `Platform` method. Initialisation
for all heaters is to the value of the `HEAT_ON` constant, so there will be a brief period during switch on where devices turn on until the software
establishes control and switches them off.