Version 1.19beta9
New and changed features:
- Experimental resume-after-power-failure support on Duet WiFi/Ethernet.
See
https://duet3d.com/wiki/Setting_up_to_resume_a_print_after_a_power_failure.
- Partly implemented bed levelling using multiple independent Z motors
(M671). See
https://duet3d.com/wiki/Bed_levelling_using_multiple_independent_Z_motors.
- If the G10 command is used to set the standby temperature of a tool
that is on standby, the live temperature is adjusted accordingly
- SCARA parameters configured using M669 now include X and Y bed origin
offsets
- Baby stepping is no longer cleared when you home the printer or probe
the bed
- The Y axis can now be mapped in a similar way to the X axis
- When the WiFi module is in access point mode, M122 displays the number
of connected clients (needs DuetWiFiServer 1.19beta9)
- The M305 command now uses the S parameter to set the heater name
instead of the H parameter
- The meaning of the first M669 crosstalk parameter for a SCARA printer
has changed. A zero value now means that the proximal motor does not
affect the proximal-to-distal arm angle.
- The CoreXY kinematics calculations have been changed to conform to the
way they are defined in other firmwares and at CoreXY.com. See the
important upgrade notes. The CoreXZ and CoreXYU kinematics have been
changed similarly.
- The maximum allowed target temperature for auto tuning now depends on
the configured maximum temperature for that heater
- The heater gain that provokes a warning when setting heater model
parameters with M307 or after auto tuning now depends on the configured
maximum temperature for that heater

Bug fixes:
- G2 and G3 arc movement commands didn't work when the X axis was mapped
- On an IDEX machine there was unwanted movement of the new tool after a
tool change occurred
- Duet WiFiServer 1.19-beta9: fixed a bug that sometimes caused WiFi
connection failures
- On the Duet WiFi, if the own access point parameters were configurds
but the list of remembered host access points was empty, the own access
point name appeared in the remembered list displayed by M587
- On a SCARA mnachine, sending G92 X0 Y0 caused the Duet to return
position data containing NaNs to Duet Web Control, which caused it to
disconnect

Upgrade notes:
- **Important!** On a CoreXY machine, you need to either swap the X and
Y motor connections, or set the Y axis factor to -1 in the M667 command.
Similarly for CoreXZ and CoreXYU machines.